% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2raw_text.R
\name{csv2raw_text}
\alias{csv2raw_text}
\title{Convert .csv to the raw text}
\usage{
csv2raw_text(file)
}
\arguments{
\item{file}{character vector with the name of the .csv file}
}
\value{
writes a csv file with the annotation
}
\description{
Convert .csv to the raw text
}
\author{
George Moroz <agricolamz@gmail.com>
}
