% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2TextGrid.R
\name{csv2TextGrid}
\alias{csv2TextGrid}
\title{Convert audio to the anotated .csv file}
\usage{
csv2TextGrid(name, speaker = "speaker")
}
\arguments{
\item{name}{character vector with the name of the .csv file}

\item{speaker}{character vector with the speaker name that is used as the tier name.}
}
\value{
writes a csv file with the annotation
}
\description{
Adds glosses to the glosses list and adds small capitals to glosses. Escapes strings that begins and ends with curly brackets.
}
\author{
George Moroz <agricolamz@gmail.com>
}
