% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/folder2code.R
\name{folder2code}
\alias{folder2code}
\title{Generate code for analysis of audio from the folder}
\usage{
folder2code(path = getwd(), model_path, language = "ru", speakers = "")
}
\arguments{
\item{path}{character vector with the path to the file}

\item{model_path}{character vector with the path to the downloded model.}

\item{language}{character vector with the code for the language. See \code{audio.whisper::whisper_languages()} for the list of languages.}

\item{speakers}{character vector with the speakers for TextGrid creation function.}
}
\value{
writes a csv file with the annotation
}
\description{
Since speech2text procedures can take a lot of time, it make sense to create code for analyzis rather then using loops. This function automatically parses the folder for audio and creates code for speech2text functions. It is also possible to add lines for adding TextGrid files.
}
\author{
George Moroz <agricolamz@gmail.com>
}
