% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/speech2text.R
\name{speech2text}
\alias{speech2text}
\title{Convert audio to the anotated .csv file}
\usage{
speech2text(audio, output_name = "output", model_path, language = "ru")
}
\arguments{
\item{audio}{character vector with the path to the file}

\item{output_name}{character vector with the name of the result csv file.}

\item{model_path}{character vector with the path to the downloded model.}

\item{language}{character vector with the code for the language. See \code{audio.whisper::whisper_languages()} for the list of languages.}
}
\value{
writes a csv file with the annotation
}
\description{
Adds glosses to the glosses list and adds small capitals to glosses. Escapes strings that begins and ends with curly brackets.
}
\author{
George Moroz <agricolamz@gmail.com>
}
