% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv2TextGrid.R
\name{csv2TextGrid}
\alias{csv2TextGrid}
\title{Convert .csv to the Praat TextGrid}
\usage{
csv2TextGrid(file, speaker = "speaker")
}
\arguments{
\item{file}{character vector with the name of the .csv file}

\item{speaker}{character vector with the speaker name that is used as the tier name.}
}
\value{
writes a csv file with the annotation
}
\description{
Convert .csv to the Praat TextGrid
}
\author{
George Moroz <agricolamz@gmail.com>
}
